==> 07_Final Output.Rmd <==
---
title: "CDI Pets Regression Tables"
author: "Jeffrey Berinstein"
date: "1/23/2019"
output:
  word_document: default
  html_document:
    df_print: paged
---

```{r setup, include=FALSE}
#knitr::opts_chunk$set(echo = TRUE)
library(leaps)
library(foreign)
library(survival)
library(moderndive)
library(tidyverse)
library(here)
library(broom)
library(flextable)
library(officer)
library(purrr)
library(knitr)
library(tidyverse)
library(msm)
library(arsenal)
library(survival)
library(foreign)
library(dplyr)
library(Hmisc)
library(tidyverse)
library(here)
library(corrr)
library(corrplot)
library(finalfit)
library(EValue)

#TableOne Stuff begins

mtable <- read_csv('02.2_Final_table_with_meds.csv')


#Convert gender into factor F=1, M=2

mtable$gender<-factor(mtable$gender, levels=c("F","M"), labels = c("Female", "Male"))

#Convert Race into factor with 1=white, 2= black, 3= Hispanic, 4=asian, 5=American Indian and Alaska Native
#6=Other"7=unknown

mtable$race<-factor(mtable$race, levels = c("white", "black", "Hispanic", "asian", 
                                        "American Indian and Alaska Native", "other","unknown"), 
                        labels=c("White", "Black", "Hispanic", "Asian", 
                      "American Indian and Alaska Native", "Other","Unknown"))

mtable$antibiotics3mo<-factor(mtable$antibiotics3mo, levels=c("0","1"), labels = c("0", "Antibiotic use"))
mtable$ppi<-factor(mtable$ppi, levels=c("0","1"), labels = c("0", "PPI use"))
mtable$h2ra<-factor(mtable$h2ra, levels=c("0","1"), labels = c("0", "H2RA use"))
mtable$any_all_med<-factor(mtable$any_all_med, levels=c("0","1"), labels = c("0", "Nasal Decongestant use"))
mtable$reccurent_cdi<-factor(mtable$reccurent_cdi, levels=c("0","1"), labels = c("0", "Prior CDI"))
mtable$dog<-factor(mtable$dog, levels=c("0","1"), labels = c("0", "Dog Exposure"))
mtable$cat<-factor(mtable$cat, levels=c("0","1"), labels = c("0", "Cat Exposure"))
mtable$hc_facility3<-factor(mtable$hc_facility3, levels=c("0","1"), labels = c("0", "SAR/SNF"))
mtable$hospital3<-factor(mtable$hospital3, levels=c("0","1"), labels = c("0", "Hospitalization"))

attr(mtable$gender,'label')  <- 'Gender'
attr(mtable$race,'label')  <- 'Race'
attr(mtable$antibiotics3mo,'label')  <- 'Antibiotic'
attr(mtable$ppi,'label')  <- 'PPI'
attr(mtable$h2ra,'label')  <- 'H2RA'
attr(mtable$hospital3,'label')  <- 'Hospitalization'
attr(mtable$hc_facility3,'label')  <- 'SAR/SNF'
attr(mtable$adl_total,'label')  <- 'Total ADL Score'
attr(mtable$dog,'label')  <- 'Dog Exposure'
attr(mtable$cat,'label')  <- 'Cat Exposure'
attr(mtable$any_all_med,'label')  <- 'Any Decongenstant'
attr(mtable$reccurent_cdi,'label') <- 'Prior CDI'

#Create Table one, stratified according to CDI status
tab1 <- tableby(cdi_status ~ gender + age + race + reccurent_cdi + antibiotics3mo +
         ppi +h2ra + any_all_med+hospital3+ hc_facility3+ adl_total + dog +cat, data=mtable)

#TableOne Stuff Ends

table_strata <- read_csv('04p_table_strata.csv')

table_strata$pet <- case_when(table_strata$dog+
                                table_strata$cat==0 ~ 0,
                              TRUE ~ 1)
table_strata$pet_allerg <- case_when(table_strata$dog_allerg+
                                table_strata$cat_allerg==0 ~ 0,
                              TRUE ~ 1)
table_strata$hc_hosp <- case_when(table_strata$hc_facility3+
                                       table_strata$hospital3==0 ~ 0,
                                     TRUE ~ 1)
table_strata$any_outside_pet <- case_when(table_strata$dog_outside+
                                       table_strata$cat_outside==0 ~ 0,
                                     TRUE ~ 1)

table_strata1<- table_strata %>% 
  select(-c(study_num, cdi))

##Correlogram

table0 <- read_csv('04p_table_strata.csv')

#Remove MC_stable as no one was not it (1 that was removed)
table<-table0 %>% 
  select(-mc_stable)

#Reaarange so all medications are close

table1<-table[c(1:19, 30:36, 20:29)]

#Temporarily remove study_num and cdi,
# and non-numeric variables
table_corr<- table1 %>% 
  select(-c(study_num, cdi, gender, race))
str(table_corr)

colnames(table_corr) <- c("CDI Status", "Prior CDI", "Patient Age", "Frequent Restaurant Visits", "High Dessert Intake", "High Meat Intake", 
                          "High Salad Intake", "Frequent Red Wine Intake", "Dairy Intake", "Vitamins", "Probiotic", "Antibiotic", 
                          "PPI",  "H2RA", "Any Acid Suppresion", "Ipratropium", "Antihistamine", 
                          "Intranasal Corticosteroids", "Intranasal Vasoconstrictor", "Antileukotriene", "Intransal Atropine",
                          "Any Decongestant", "Healthcare Occupation", "Hospitalization", "SAR/SNF", "Total ADL Score ",
                          "Dog Allergy", "Cat Allergy", "Any Dog", "Outside Dog", "Any Cat", "Outside Cat")
                        
col <- colorRampPalette(c("#BB4444", "#EE9988", "#FFFFFF", "#77AADD", "#4477AA"))

p.mat <- cor.mtest(table_corr)$p

```

```{r TableOne, results='asis'}


summary(tab1, digits=3, digits.p =2, digits.pct =1) 

```

## Stratification Results
**Male**
Median Difference: 1.2, 
Min Difference: 0
Max Difference: 5.2

**Female**
Median Difference: 1.0, 
Min Difference: 0
Max Difference: 4.1

## Corelogram

```{r correlogram, echo=FALSE}

corrplot(cor(table_corr), method = "color", col = col(200),
         type = "upper", order = "AOE", number.cex = .6,
         #addCoef.col = "black", # Add coefficient of correlation
         tl.col = "black", tl.srt = 90, # Text label color and rotation
         # Combine with significance
         p.mat = p.mat, sig.level = 0.05, pch.cex = 1, insig = "label_sig", 
         # hide correlation coefficient on the principal diagonal
         diag = FALSE)
```

## Bivariate Conditional Regression


```{r Bivariate Regression, echo=FALSE}


full_df <-  table_strata1  #1
 predictors_df <- table_strata1 %>%  #2
   dplyr::select(-c(cdi_status, reccurent_cdi, gender, age, race, stratum, mc_stable))
 
 predictors_df %>% 
   map(~survival::clogit(full_df$cdi_status~.x, data=full_df)) %>% 
   map(tidy) -> #4
   list
 
 names(list)<- c( "Frequent Restaurant Visits","High Dessert Intake", "High Meat Intake", "High Salad Intake", "Frequent Red Wine Intake", "Dairy Intake", "Vitamins",
                 "Probiotic", "Antibiotic", "PPI",  "H2RA", "Any Acid Suppression", "Healthcare Occupation","Hospitalization", "SAR/SNF",
                 "ADL Score ", "Dog Allergy", "Cat Allergy", "Dog", "Outside Dog", "Cat", "Outside Cat", "Ipratropium", "Antihistamine",
                  "Intranasal Corticosteroids", "Intranasal Vasoconstrictor", "Antileukotriene", "Intranasal Atropine",
                  "Any Nasal Decongestant", "Any Pet", "Any Pet Allergy", "Hospitalization/SAR/SNF", "Any Outside Pet")


pred <- names(list)
list %>% 
  bind_rows() %>% #6
  filter(term != "(Intercept)") %>% 
  mutate(predictor = pred) %>% 
  mutate(odds_ratio = exp(estimate),
         LCB = exp(conf.low),
         UCB = exp(conf.high)) %>% 
  select(predictor, odds_ratio, LCB, UCB, p.value) %>% 
  arrange(p.value) -> #8
  bivariate_table_or


kable(bivariate_table_or)

```

```{r Bivariate Regression Plot, echo=FALSE}

bivariate_table_or<-bivariate_table_or %>% 
  filter(predictor!="Prior CDI")

uni_graph <-ggplot(bivariate_table_or, aes(reorder(predictor, odds_ratio))) + geom_point(aes(y=odds_ratio)) +
  geom_errorbar(aes(ymin=`LCB`, ymax=`UCB`), width = 0.8) + coord_flip() 

uni_graph + theme_bw() + theme(panel.border = element_blank(), panel.grid.major = element_blank(),
                        panel.grid.minor = element_blank(), axis.line = element_line(colour = "black")) +xlab("CDI Predictors") +ylab("Odds Ratio") +
  geom_hline(yintercept = 1, linetype = "dotted")

```


## Multi Variable Conditional Regression without interactions

### With Dogs in the Model(best model)

```{r bivariate table with pet, echo=FALSE}

res.clogit.int <- survival::clogit(cdi_status ~ 
                                   #reccurent_cdi+
                                    #restaurant3 + 
                                    #dessert3 +
                                    meat3 +
                                    salad3 +
                                    #redwine3 + 
                                    #dairy +
                                    #vitamin+ 
                                    #probiotic+
                                    #antibiotics3mo +
                                   #ppi+
                                    #h2ra +
                                     #acid_blocker_yes +
                                     #health_care +
                                     #hospital3 +
                                     #hc_facility3 +
                                     adl_total +
                                     #dog_allerg +
                                     #cat_allerg +
                                     #dog +
                                     #dog_outside +
                                     #cat +
                                     #cat_outside +
                                     #ipratropium +
                                     #antihistamine +
                                     #intranasal_cs +
                                     #intranasal_vc +
                                     #antileukotriene +
                                     #atropine_nasal +
                                     #any_all_med +
                                     pet+
                                     #pet_allerg +
                                     hc_hosp+
                                     #any_outside_pet +
                                     strata(stratum), table_strata1)

MO1 <- tidy(res.clogit.int)

ft_dog <- MO1 %>% 
  mutate(odds_ratio = exp(estimate),
         LCB = exp(conf.low),
         UCB = exp(conf.high)) %>% 
  select(term, odds_ratio, LCB, UCB, p.value) %>% 
  arrange(desc(odds_ratio))



# ft_dog$term<- c("High Meat Intake", "Hospitalization/SAR/SNF", " ADL Score","Pet")


kable(ft_dog, digits = 3)

```



```{r plot OR dog, echo=FALSE}

gdog<-ggplot(ft_dog, aes(reorder(term, odds_ratio))) + geom_point(aes(y=odds_ratio)) +
  geom_errorbar(aes(ymin=`LCB`, ymax=`UCB`), width = 0.2) + coord_flip() 

gdog + theme_bw() + theme(panel.border = element_blank(), panel.grid.major = element_blank(),
                       panel.grid.minor = element_blank(), axis.line = element_line(colour = "black")) +xlab("CDI Predictors") +
  ylab("Odds Ratio") +
  geom_hline(yintercept = 1, linetype = "dotted") 

```

### With Predicitng Pets in the Model(best model)


```{r predicting pets, echo=FALSE}

res.clogit.int1 <- survival::clogit(pet~
                                      #cdi_status+
                                   #reccurent_cdi+
                                     #age+
                                    #restaurant3 + 
                                    #dessert3 +
                                    #meat3 +
                                    salad3 +
                                    #redwine3 + 
                                    #dairy +
                                    #vitamin+ 
                                    #probiotic+
                                    #antibiotics3mo +
                                   #ppi+
                                    #h2ra +
                                     #acid_blocker_yes +
                                     #health_care +
                                  #hospital3 +
                                     #hc_facility3 +
                                     adl_total +
                                     #dog_allerg +
                                     #cat_allerg +
                                     #ipratropium +
                                     #antihistamine +
                                     #intranasal_cs +
                                     #intranasal_vc +
                                     #antileukotriene +
                                     #atropine_nasal +
                                     #any_all_med +
                                     #pet_allerg +
                                     #hc_hosp+
                                     #any_outside_pet +
                                     strata(stratum), table_strata1)


MO2 <- tidy(res.clogit.int1)

ft_pet <- MO2 %>% 
  mutate(odds_ratio = exp(estimate),
         LCB = exp(conf.low),
         UCB = exp(conf.high)) %>% 
  select(term, odds_ratio, LCB, UCB, p.value) %>% 
  arrange(desc(odds_ratio))



ft_pet$term<- c("ADL Score","High Salad Intake")


kable(ft_pet, digits = 3)


res.clogit.int1_pro <- survival::clogit(probiotic~
                                      #cdi_status+
                                   reccurent_cdi+
                                     strata(stratum), table_strata1)


MO2_pro <- tidy(res.clogit.int1_pro)

ft_pro <- MO2_pro %>% 
  mutate(odds_ratio = exp(estimate),
         LCB = exp(conf.low),
         UCB = exp(conf.high)) %>% 
  select(term, odds_ratio, LCB, UCB, p.value) %>% 
  arrange(desc(odds_ratio))



# ft_pet$term<- c("ADL Score","High Salad Intake")


kable(ft_pro, digits = 5)

```


### Without probiotics in the model (this will be the final model reported)

```{r bivariate table no intereactions, echo=FALSE}

res.clogit4.int <- survival::clogit(cdi_status ~  
                #probiotic + 
                  meat3 + #hospital3+
                  hc_hosp + 
                  #hc_facility3 +
                  I(adl_total/3) + #vitamin +
                  cat_allerg +  
                  salad3 + #health_care + 
                  #vitamin+
                  #h2ra +
                   #antibiotics3mo + 
                 strata(stratum), 
                    table_strata1)

model_output1 <- tidy(res.clogit4.int)

summary(res.clogit4.int)

ft1 <- model_output1 %>% 
  mutate(odds_ratio = exp(estimate),
         LCB = exp(conf.low),
         UCB = exp(conf.high)) %>% 
  select(term, odds_ratio, LCB, UCB, p.value) %>% 
  arrange(desc(odds_ratio))


ft1$term<- c("High Meat Intake",
             "Cat Allergy",
             "Hospitalization/SAR/SNF",
                        "ADL Score",  "High Salad Intake")

kable(ft1,digits=3, digits.p =2, digits.pct =1)


```

```{r multivariate plot no intereactions, echo=FALSE}


g1<-ggplot(ft1, aes(reorder(term, odds_ratio))) + geom_point(aes(y=odds_ratio)) +
  geom_errorbar(aes(ymin=`LCB`, ymax=`UCB`), width = 0.2) + coord_flip() 

g1 + theme_bw() + theme(panel.border = element_blank(), panel.grid.major = element_blank(),
                       panel.grid.minor = element_blank(), axis.line = element_line(colour = "black")) +xlab("CDI Predictors") +
  ylab("Odds Ratio") +
  geom_hline(yintercept = 1, linetype = "dotted") 
  
 
```

### With probiotics in the model

```{r multivariate table w/ probiotic, echo=FALSE}

res.clogit.int_pro <- survival::clogit(cdi_status ~ 
                                   #reccurent_cdi+
                                    #restaurant3 + 
                                    #dessert3 +
                                    meat3 +
                                    #salad3 +
                                    #redwine3 + 
                                    #dairy +
                                    #vitamin+ 
                                    probiotic+
                                    #antibiotics3mo +
                                   #ppi+
                                    #h2ra +
                                     #acid_blocker_yes +
                                     #health_care +
                                     #hospital3 +
                                     #hc_facility3 +
                                     adl_total +
                                     #dog_allerg +
                                     #cat_allerg +
                                     #dog +
                                     #dog_outside +
                                     #cat +
                                     #cat_outside +
                                     #ipratropium +
                                     #antihistamine +
                                     #intranasal_cs +
                                     #intranasal_vc +
                                     #antileukotriene +
                                     #atropine_nasal +
                                     #any_all_med +
                                     #pet+
                                     #pet_allerg +
                                     hc_hosp+
                                     #any_outside_pet +
                                     strata(stratum), table_strata1)


model_output_pro <- tidy(res.clogit.int_pro)

ft_pro <- model_output_pro %>% 
  mutate(odds_ratio = exp(estimate),
         LCB = exp(conf.low),
         UCB = exp(conf.high)) %>% 
  select(term, odds_ratio, LCB, UCB, p.value) %>% 
  arrange(desc(odds_ratio))


ft_pro$term<- c("Probiotic", "High Meat Intake" , "Hospitalization/SAR/SNF", "Total ADL Score")

kable(ft_pro, digits = 4)
```


```{r multivariate plot with probiotics, echo=FALSE}

g_pro<-ggplot(ft_pro, aes(reorder(term, odds_ratio))) + geom_point(aes(y=odds_ratio)) +
  geom_errorbar(aes(ymin=`LCB`, ymax=`UCB`), width = 0.2) + coord_flip() 

g_pro + theme_bw() + theme(panel.border = element_blank(), panel.grid.major = element_blank(),
                       panel.grid.minor = element_blank(), axis.line = element_line(colour = "black")) +xlab("CDI Predictors") +
  ylab("Odds Ratio") +
  geom_hline(yintercept = 1, linetype = "dotted") 

```

## 2 x2 table for CDI status vs Probiotic exposure
```{R build 2x2 tabletable with n, echo =FALSE}

explanatory = "probiotic"
dependent= "cdi_status"

table_strata2<- table_strata1

table_strata2$cdi_status <- as.factor(table_strata2$cdi_status)
table_strata2$probiotic <- as.factor(table_strata2$probiotic) 

table_strata2 %>% 
  summary_factorlist(dependent, explanatory,
  p=TRUE, add_dependent_label=TRUE) %>% kable()

```
### Calculating an E-value of probiotics

The E- value measures the strength of association that an unmeasured confounder would have to have in order to cancel out an observed relationship.

E-Value for probiotics based on EValue Package. Note they calculate RR as sqrt of OR

```{r Evalue, echo=FALSE}

evalues.OR(2.819, 1.532,5.189, rare = FALSE, true = 1) %>% kable()

```


### E-Value based on manual calculation of Relative Risk

- Convert OR to RR with the following formula :RR = OR / (1 – p + (p x OR))
- p=outcome of interest in the non-exposed group (i.e CDI+ priobiotic unexposed)

```{r, table with RR and Evalue, echo=FALSE}

ft_pro1<- ft_pro %>% 
  mutate(risk_ratio= (odds_ratio/((1 - 0.466) + (0.466 * odds_ratio))), LCB_OR=(LCB/((1 - 0.466) + (0.466 * LCB))),  
                      UCB_OR=(UCB/((1 - 0.466) + (0.466 * UCB))))
  
kable(ft_pro1)


evalues.RR(1.525820, 1.227670, 1.757770) %>% kable()
  
```
